Integrating dockerized jenkins in jazz installer
Former-commit-id: afc78686e6adb839f45e05f2a301f68c483f0b43

+++ installscripts/dockerfiles/jenkins/launch_jenkins_docker.sh
@@ -1,6 +1,5 @@
 #Installing Docker-ce on centos7
 sudo yum check-update
-sudo yum remove docker docker-common docker-selinux docker-engine -y
 sudo yum install -y yum-utils device-mapper-persistent-data lvm2
 sudo yum-config-manager --add-repo https://download.docker.com/linux/centos/docker-ce.repo
 sudo yum install docker-ce -y
@@ -10,22 +9,49 @@ sudo systemctl enable docker
 sudo usermod -aG docker $(whoami)
 
 #Configuring and pulling the jenkins-docker from ECR
-sudo docker volume rm jenkins-volume
+sudo docker volume inspect jenkins-volume
+if [ $? == 0 ] ; then
+  sudo docker volume rm jenkins-volume
+fi
+
+region=`grep region ~/jazz-installer/installscripts/jazz-terraform-unix-noinstances/variables.tf  | cut -d' ' -f 9 | tr -d '"'`
 sudo docker volume create jenkins-volume
-sudo aws ecr get-login --no-include-email --region us-east-1 > ecr_login_script
+aws ecr get-login --no-include-email --region $region > ecr_login_script
 sudo bash ecr_login_script && rm -f ecr_login_script
-sudo docker pull 108206174331.dkr.ecr.us-east-1.amazonaws.com/jazz-oss-repo:jenkins-docker
-sudo docker stop jenkins-server
-sudo docker rm jenkins-server
-sudo docker run -dt -p 2200:2200  -p 8080:8080 --name=jenkins-server --mount source=jenkins-volume,destination=/var/lib/jenkins 108206174331.dkr.ecr.us-east-1.amazonaws.com/jazz-oss-repo:jenkins-docker
-sudo docker cp jenkins-server:/root/.ssh/id_rsa private_key.pem
+sudo docker pull 108206174331.dkr.ecr.us-east-1.amazonaws.com/jazz-oss:jenkins
+
+sudo docker ps -a | grep -i jenkins-server
+if [ $? == 0 ] ; then
+  sudo docker stop jenkins-server
+  sudo docker rm jenkins-server
+fi
+sudo docker run -dt -p 2200:2200 -p 8080:8080 --name=jenkins-server --mount source=jenkins-volume,destination=/var/lib/jenkins 108206174331.dkr.ecr.us-east-1.amazonaws.com/jazz-oss:jenkins
+sudo docker cp jenkins-server:/root/.ssh/id_rsa ~/jenkinskey.pem
+sudo chmod +r ~/jenkinskey.pem
+echo "==============>"
+echo "You will see the jenkins server login details after the stack creation!"
 echo "Waiting for Jenkins Server to be ready..."
-sleep 8
+sleep 20
 ip=`curl -sL http://169.254.169.254/latest/meta-data/public-ipv4`
 initialPassword=`sudo cat /var/lib/docker/volumes/jenkins-volume/_data/secrets/initialAdminPassword`
-echo "============>"
-echo "Jenkins Docker Details:"
-echo "PublicIp: http://$ip:8080"
-echo "Username: admin"
-echo "Password: $initialPassword"
-echo "============>"
+mac=`curl -sL http://169.254.169.254/latest/meta-data/network/interfaces/macs`
+security_groups=`curl -sL http://169.254.169.254/latest/meta-data/network/interfaces/macs/${mac%/}/security-group-ids`
+subnet_id=`curl -sL http://169.254.169.254/latest/meta-data/network/interfaces/macs/${mac%/}/subnet-id`
+#values to be passed
+jenkins_server_elb="$ip:8080"
+jenkins_username="admin"
+jenkins_passwd="$initialPassword"
+jenkins_server_public_ip="$ip"
+jenkins_server_ssh_login="root"
+jenkins_server_ssh_port="2200"
+jenkins_server_security_group="$security_groups"
+jenkins_server_subnet="$subnet_id"
+
+echo "$jenkins_server_elb" > docker_jenkins_vars
+echo "$jenkins_username" >> docker_jenkins_vars
+echo "$jenkins_passwd" >> docker_jenkins_vars
+echo "$jenkins_server_public_ip" >> docker_jenkins_vars
+echo "$jenkins_server_ssh_login" >> docker_jenkins_vars
+echo "$jenkins_server_ssh_port" >> docker_jenkins_vars
+echo "$jenkins_server_security_group" >> docker_jenkins_vars
+echo "$jenkins_server_subnet" >> docker_jenkins_vars
